refactor: simplify session recording v2 file names
