+++ plugin-server/src/main/ingestion-queues/session-recording-v2/consumer.ts
@@ -12,8 +12,8 @@ import { PluginServerService, PluginsServerConfig, TeamId, ValueMatcher } from '
 import { status } from '../../../utils/status'
 import { runInstrumentedFunction } from '../../utils'
 import { addSentryBreadcrumbsEventListeners } from '../kafka-metrics'
-import { KafkaMetrics } from './kafka/kafka-metrics'
-import { KafkaParser } from './kafka/kafka-parser'
+import { KafkaMetrics } from './kafka/metrics'
+import { KafkaParser } from './kafka/parser'
 import { SessionRecordingMetrics } from './metrics'
 import { PromiseScheduler } from './promise-scheduler'
 import { IncomingRecordingMessage } from './types'
